+++ resdk/resolwe.py
@@ -13,6 +13,7 @@ Resolwe
 """
 from __future__ import absolute_import, division, print_function
 
+import copy
 import os
 import re
 import uuid
@@ -305,6 +306,7 @@ class Resolwe(object):
             raise ValueError("Unexpected behaviour at get process with slug {}".format(slug))
 
         # Pre-process inputs
+        input = copy.deepcopy(input)  # leave original intact
         try:
             for schema, fields in iterate_fields(input, process['input_schema']):
                 field_name = schema['name']